Enable CONFIG_NAMESPACES everywhere (rhbz 907576)
So apparently we've been enabling CONFIG_NAMESPACES in Fedora kernels for
a very long time even though it was never explicitly set.  The kconfig entry
for it looks like it should depend on CONFIG_EXPERT (which isn't set), but
either the fact that we've had CONFIG_NET_NS, etc enabled explicitly or the
fact that it defaults to !EXPERT seems to do something weird with menuconfig
and enables it in the final config.  Whatever, just turn it back on.  It
seems to break rtkit-daemon for some reason.