Update PatchList
Try and keep this up to date as we add patches.  Should make my life easier
every month when I sent out the fedora kernel report.

The alternative is to list the information in each patch itself, which we
could do in place of this (or even in addition to it).  Opinions?  (As if
anyone will read this and give me opinions.)