Reduce buildid definition clutter
The buildid mechanism is still used by a nubmer of people, but it's
unnecessarily complicated in the spec.  We don't need to do all the
manipulations that are current done just to allow people to easily
mark their builds as different.  Just leave the commented out
definition and include it in pkg_release.