Split kernel into kernel-core and kernel-drivers subpackages
This creates kernel-core and kernel-drivers subpackages.  The kernel package
remains as a meta-package the requires both of the subpackages.  This allows
most installs to continue on as-is with upgrades working.

The contents of the kernel-core and kernel-drivers subpackages are determined
at build time through the filter-modules.sh script.  This script "removes"
pre-defined subsystems and modules and generates a filelist for the %files
section of each subpackage.  The contents of each are per-arch, with arch
override files taken into account.  This allows us to make the split useful
for varying arches.