Remove with_extra switch
The with_extra switch was originally added with the addition of the
modules-extra subpackage.  However, it never really worked.  If you build
with it disabled, the subpackage is omitted but the modules that would be
included in it would still be built.  That can lead to errors further in the
build when we're moving things around in the -core and -modules subpackages.

Since it doesn't actually save build time at all, and it leads to build
failures for no real gain, drop it.