Fix 3.19-rc6 builds with patch-2.7.3
Upstream patch decided that relative symlinks are dangerous to create via
patch, so it refuses to do so.  To get the kernel to build, we remove the
hunk in patch-3.19-rc6.xz that was attempting to create one, and do it
manually in the spec file.  The spec change should go away when 3.19 final
is out.