Linux v4.0-rc1
- CVE-2015-0275 ext4: fallocate zero range page size > block size BUG (rhbz 1193907 1195178)
- Disable debugging options.

Yay for major version bumps :\.

We grab the full rc1 tarball instead of just the patch to make this way less
messy in the spec.  When 4.0 final is released, we'll need to undo a few
specific hacks.  Namely, we need to redefine:

upstream_sublevel
kversion
Source0

back to the standard definitions.

Thanks to Kyle for figuring this out 3 years ago.