kernel.spec: no more files in /boot
This patch is to install everything to /lib/modules/$KERNEL_VERSION
 and only have %ghost files for /boot files.

Installation in %posttrans with kernel-install should work already.

F20: systemd >= 208-31
F21: systemd >= 216-21
F >= 22: systemd >= 215-12

The main target is to have all rpms install only to /usr, so in the end
we would have a self contained installation in /usr.

As of now, we still have to cope with /var and /etc, but you have to
start somewhere.