Make the git snapshot generator work for -git1 patches
Before it would just error out because it was only looking at the patch-
listings in the sources file.  So now we let it do that, and if the VER
variable is null after we just assume it's the first git snapshot of the
merge window and figure out the version from the linux- tarball.

Why I didn't fix this sooner, I have no idea.  Probably because it only
happens 4 times a year and laziness wins the day.