Refresh kdbus.patch
Update it so it isn't a hodge podge of random patches.  Also fix it so things
after the 'kdbus: selftests: add more name registry tests' commit show up as
individual commits and not lumped into that one.  No idea why that was
happening.

Finally, add another patch from David to increase the limit on matches