Disable hibernation for powerpc (rhbz 1267395)
A recent binutils update (correctly) started checking for the
number of arguments to tlbie. This broke some tlb invalidation
assembly that was invoked during the hibernation path and
created compilation errorrs. Unfortunately, fixing it is not
an easy task because nobody is completely sure if it actually
works. Since this is preventing anything from powerpc from being
built at all, disable hibernation for now. If/when this gets
fixed upstream this feature can be turned back on.