Add kernel-cross-headers subpackage
The kernel-headers package it primarily used in Fedora to build the
toolchain components.  This works fine for native builds, but we do
not have cross-architecture headers available.  That means that the
cross compiler toolchains we have can only be built without headers
and therefore are only useful for building the kernel.  Given that
Fedora does not allow cross-compiling of it's packages, that hasn't
been a problem to date.

However, the toolchains might be used by application developers to
build their own software.  Providing the ability to build cross
toolchains that offer glibc as well may benefit these users.  This
provides a kernel-cross-headers package that will at least allow
this possibility.