Add patches to fix boot on 64K page aarch64 devices (eg Seattle)
Add the following series that in theory fixes boot on aarch64. It's ACKed
and has pull reqs for upstream so in theory upstream should have tested
this, although I have my doubts. If it is in fact crack just revert and
I'll go and beat ^W work something else out.