kernel.spec: Fix build dep on libelf-devel
The dependency on libelf-devel is wrapped in "%if %{with_perf}".
However, if you're building without perf (as fast-build.sh does) then
you don't have this installed and the build will complain:

Makefile:1037: "Cannot use CONFIG_STACK_VALIDATION, please install
libelf-dev, libelf-devel or elfutils-libelf-devel"

This just lifts the dependency out of the %if section.