Add scripts for automatically generating builds
Many of the steps for day-to-day updates still require a lot of
manual steps which can lead to typos. Add a series of scripts to
make automate the manual steps and hopefully reduce errors. These
are a work in progress because having them in tree makes it easier
to work out issues.