Don't list /lib/modules/$KernelVer/kernel twice
For quite some time I see warnings like
    warning: File listed twice: /lib/modules/[...]/kernel

when building the kernel-core rpm. This is because that directory is included
both in module-dirs.list (and therefor in, basically, kernel-core.list) and in
kernel_variant_files().

Tweak the find command that generates module-dirs.list so that this directory
is not included in that list anymore.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>