Add script to check config generation
The kernel configuration generation currently checks to make sure
every option is defined with listnewconfig. It does not check that
each option is the same as listnewconfig. This can lead to odd
situations where the Fedora configuration does not match what's
actually present in the generated config. Add a script to check
for these kinds of changes.

Based on work done by Miguel Flores Silverio <floresmigu3l@gmail.com>