Reorder patch numbers to a rough scheme
A little while ago a short lived scheme was tried to maintain kernel patches
outside of the Fedora spec in a dedicated kernel source repo. The patch numbers
were then auto-gen to add the various patches to the spec. The current numbering
in the 400-500 range was a result of that. Now it's not used lets move back to
something a little more organised.