Update debuginfo generation
Once upon a time, the kernel needed a lot of special handling to
generate proper debuginfo as the kernel was ahead in technology. These
days, rpm has improved debuginfo support. The kernel has not kept up
with this and it's forward looking calls are now out of date. Switch to
more standard invocations of debuginfo calls.